refactor: import auth dependency from its defining module
projects.py and items.py imported get_current_user from cost_query_pro.api.auth,
which merely re-imports it from cost_query_pro.core.security. Under
--strict (no_implicit_reexport) that is not a valid export.

Import from core.security, where the function is actually defined, rather
than adding an __all__ re-export to api.auth to prop up the indirection.

Runtime behavior is unchanged — it was always the same function object, so
FastAPI dependency resolution and the existing test overrides (which already
key on core.security.get_current_user) are unaffected.

Also drop two obsolete arg-type ignores in services/ingestion.py. The parsed
rows are dict[str, Any], so float(row[...]) and int(row[...]) type-check on
their own; the ignores predate that signature and are not replaced.
