+++ src/cost_query_pro/api/projects.py
@@ -5,8 +5,8 @@ from typing import List, Optional
 from fastapi import APIRouter, Depends, Query, Response, status
 from sqlalchemy.orm import Session
 
-from cost_query_pro.api.auth import get_current_user
 from cost_query_pro.core.errors import AppError
+from cost_query_pro.core.security import get_current_user
 from cost_query_pro.db.session import get_db
 from cost_query_pro.models import Item, Project
 from cost_query_pro.models.user import User as DBUser
